LayoutTests/platform/blackberry/fast/dom/Window/window-early-properties-expected.txt

CONSOLE MESSAGE: line 2: TypeError: Result of expression 'window.opener.concludeTest' [undefined] is not a function.
FAIL: Timed out waiting for notifyDone to be called
Opened Window
